feat: Stop writing conditional in-lined HTML/JS
The core goal of a template-based website is to "build" pages from
pieces, not write pages in-lined. To move the site back in that
direction we are moving all in-lined html/javascript to external
"partial" assets.  We can in-line them using Hugo's Resources.Get later
(and clean them up).  This makes the final structure dramatically easier
to modify/edit, allows us to fast experiment with alternate partials,
and cleans up the visual clutter substantially.